+++ packaging/homebrew/moonlab.rb
@@ -16,43 +16,38 @@ class Moonlab < Formula
   depends_on "cmake" => :build
   depends_on "ninja" => :build
   depends_on "libomp"
+  depends_on "openssl@3"
 
   def install
-    # Set OpenMP paths for Apple Silicon
     libomp = Formula["libomp"]
-    ENV["LDFLAGS"] = "-L#{libomp.opt_lib} #{ENV["LDFLAGS"]}"
-    ENV["CPPFLAGS"] = "-I#{libomp.opt_include} #{ENV["CPPFLAGS"]}"
+    openssl = Formula["openssl@3"]
 
-    # Build using Make (primary build system)
-    system "make", "clean"
-    system "make"
+    args = std_cmake_args + %W[
+      -G Ninja
+      -DOPENSSL_ROOT_DIR=#{openssl.opt_prefix}
+      -DQSIM_BUILD_TESTS=OFF
+      -DQSIM_BUILD_EXAMPLES=OFF
+      -DQSIM_BUILD_BENCHMARKS=OFF
+    ]
 
-    # Build tests
-    system "make", "tests", "unit_tests"
+    ENV.append "LDFLAGS", "-L#{libomp.opt_lib}"
+    ENV.append "CPPFLAGS", "-I#{libomp.opt_include}"
 
-    # Run tests to verify build
-    system "make", "test_unit"
-
-    # Install library files
-    lib.install "lib/libquantum_sim.a"
-
-    # Install headers
-    include.install Dir["include/*.h"]
-
-    # Also install to lib/moonlab for namespacing
-    (lib/"moonlab").mkpath
-    (lib/"moonlab").install "lib/libquantum_sim.a"
+    system "cmake", "-S", ".", "-B", "build", *args
+    system "cmake", "--build", "build", "--target", "quantumsim", "moonlab-control-server"
+    system "cmake", "--install", "build"
   end
 
   def caveats
     <<~EOS
       Moonlab Quantum Simulator has been installed!
 
-      The static library is available at:
-        #{lib}/libquantum_sim.a
+      CMake consumers can use:
+        find_package(quantumsim CONFIG REQUIRED)
+        target_link_libraries(app PRIVATE quantumsim::quantumsim)
 
-      To use in your project, link with:
-        -L#{lib} -lquantum_sim -lomp
+      C ABI headers are available under:
+        #{include}/moonlab
 
       For Python bindings:
         pip install moonlab
@@ -62,7 +57,26 @@ class Moonlab < Formula
   end
 
   test do
-    # Test that library was built correctly
-    assert_predicate lib/"libquantum_sim.a", :exist?
+    assert_predicate lib/"libquantumsim.dylib", :exist? if OS.mac?
+    assert_predicate include/"moonlab/moonlab_export.h", :exist?
+    assert_predicate lib/"cmake/quantumsim/quantumsim-config.cmake", :exist?
+
+    (testpath/"consumer.c").write <<~C
+      #include <moonlab/moonlab_export.h>
+
+      int main(void) {
+          int major = -1;
+          int minor = -1;
+          int patch = -1;
+          moonlab_abi_version(&major, &minor, &patch);
+          return major < 0 || minor < 0 || patch < 0;
+      }
+    C
+
+    system ENV.cc, "consumer.c", "-I#{include}", "-L#{lib}", "-lquantumsim",
+                   "-o", "consumer"
+    ENV.prepend_path "DYLD_LIBRARY_PATH", lib.to_s if OS.mac?
+    ENV.prepend_path "LD_LIBRARY_PATH", lib.to_s if OS.linux?
+    system "./consumer"
   end
 end